bfp-simulation-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
stable-baselines3[extra]
sb3-contrib
gymnasium
opencv-python-headless
numpy
Pillow
pydantic
psycopg2-binary
gunicorn
nltk
tensorflow
keras
google-generativeai
